// backend/src/enums/chat.enums.ts
export enum ChatEnums {
  ASSOCIATE = 'ASSOCIATE',
  NEW_USER_MESSAGE = 'NEW_USER_MESSAGE',
  NEW_ROOM_MESSAGE = 'NEW_ROOM_MESSAGE',
  JOIN_USER_TO_THE_ROOM = 'JOIN_USER_TO_THE_ROOM',
  USER_LEFT_THE_ROOM = 'USER_LEFT_THE_ROOM',
  USER_IS_TYPING = 'USER_IS_TYPING',
  USER_IN_ROOM_IS_TYPING = 'USER_IN_ROOM_IS_TYPING',
  USER_READ_MESSAGE = 'USER_READ_MESSAGE',
  USER_DISCONNECTED = 'USER_DISCONNECTED',
  USER_CONNECTED = 'USER_CONNECTED',
}